Refactor naming etc
