+++ front-end/src/App.css
@@ -6,14 +6,3 @@
   display: flex;
   justify-content: center;
 }
-
-.particles {
-  position: center;
-  width: 100%;
-  height: 100%;
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  z-index: -1;
-}
