+++ front-end/src/App.js
@@ -2,19 +2,15 @@ import React, { Component } from 'react'
 // import './App.css';
 import {Switch, Route} from 'react-router-dom'
 import Stripe from './components/stripe'
-import LoginForm from './components/loginform'
-import LandingPage from './components/landingpage.js'
+import HomePage from './components/homepage.js'
 import Navbar from './components/navbar'
-import RegistrationForm from './components/registrationform'
 
 class App extends Component {
   render() {
     return (
       <div>
         <Switch>
-          <Route exact path="/reg" component={RegistrationForm} />
-          <Route exact path="/login" component={LoginForm} />
-          <Route exact path="/" component={LandingPage} />
+          <Route exact path="/" component={HomePage} />
           <Route exact path="/stripe" component={Stripe}/>
           <Route exact path="/navbar" component={Navbar}/>
         </Switch>
